refactor: extract FormNoAccess component
